+++ lib/lightning/auditing/audit.ex
@@ -46,7 +46,7 @@ defmodule Lightning.Auditing.Audit do
 
     event_signature =
       quote do
-        def event(event, item_id, actor_id, changes \\ %{}, metadata \\ %{})
+        def event(event, item_id, actor, changes \\ %{}, metadata \\ %{})
       end
 
     event_log_functions =
@@ -54,13 +54,13 @@ defmodule Lightning.Auditing.Audit do
         quote do
           # Output:
           #
-          # def event(item_type, "foo_event", item_id, actor_id, changes) do
-          #   Lightning.Audit.event(item_type, "foo_event", item_id, actor_id, changes)
+          # def event(item_type, "foo_event", item_id, actor, changes, metadata) do
+          #   Lightning.Audit.event(item_type, "foo_event", item_id, actor, changes, metadata)
           # end
           def event(
                 unquote(event_name),
                 item_id,
-                actor_id,
+                actor,
                 changes,
                 metadata
               ) do
@@ -68,7 +68,7 @@ defmodule Lightning.Auditing.Audit do
               unquote(item),
               unquote(event_name),
               item_id,
-              actor_id,
+              actor,
               changes,
               metadata,
               &update_changes/1
@@ -125,7 +125,11 @@ defmodule Lightning.Auditing.Audit do
     field :item_id, Ecto.UUID
     embeds_one :changes, Changes
     field :actor_id, Ecto.UUID
-    field :actor, :map, virtual: true
+
+    field :actor_type, Ecto.Enum,
+      values: [:project_repo_connection, :trigger, :user]
+
+    field :actor_display, :string, virtual: true
     field :metadata, :map, default: %{}
 
     timestamps(updated_at: false)
@@ -138,13 +142,23 @@ defmodule Lightning.Auditing.Audit do
         update_changes_fun \\ fn x -> x end
       ) do
     audit
-    |> cast(attrs, [:event, :item_id, :actor_id, :item_type, :metadata])
+    |> cast(
+      attrs,
+      [
+        :event,
+        :item_id,
+        :actor_id,
+        :actor_type,
+        :item_type,
+        :metadata
+      ]
+    )
     |> cast_embed(:changes,
       with: fn schema, changes ->
         Changes.changeset(schema, changes, update_changes_fun)
       end
     )
-    |> validate_required([:event, :actor_id])
+    |> validate_required([:event, :actor_id, :actor_type])
   end
 
   @doc """
@@ -170,7 +184,7 @@ defmodule Lightning.Auditing.Audit do
 
   @doc """
   Creates a `changeset` for the `event` identified by `item_id` and caused
-  by `actor_id`.
+  by `actor`.
 
   The given `changes` can be either `nil`, `Ecto.Changeset`, struct or map.
 
@@ -181,7 +195,9 @@ defmodule Lightning.Auditing.Audit do
           String.t(),
           String.t(),
           Ecto.UUID.t(),
-          Ecto.UUID.t(),
+          Lightning.Accounts.User.t()
+          | Lightning.VersionControl.ProjectRepoConnection.t()
+          | Lightning.Workflows.Trigger.t(),
           Ecto.Changeset.t() | map() | nil,
           map(),
           update_changes_fun :: (map() -> map())
@@ -192,7 +208,7 @@ defmodule Lightning.Auditing.Audit do
         item_type,
         event,
         item_id,
-        actor_id,
+        actor,
         changes \\ %{},
         metadata \\ %{},
         update_fun \\ fn x -> x end
@@ -215,7 +231,7 @@ defmodule Lightning.Auditing.Audit do
         item_type,
         event,
         item_id,
-        actor_id,
+        actor,
         %Ecto.Changeset{data: %subject_schema{} = data, changes: changes},
         metadata,
         update_fun
@@ -242,20 +258,20 @@ defmodule Lightning.Auditing.Audit do
       item_type,
       event,
       item_id,
-      actor_id,
+      actor,
       changes,
       metadata,
       update_fun
     )
   end
 
-  def event(item_type, event, item_id, actor_id, changes, metadata, update_fun)
+  def event(item_type, event, item_id, actor, changes, metadata, update_fun)
       when is_map(changes) do
     audit_changeset(
       item_type,
       event,
       item_id,
-      actor_id,
+      actor,
       changes,
       metadata,
       update_fun
@@ -266,7 +282,7 @@ defmodule Lightning.Auditing.Audit do
          item_type,
          event,
          item_id,
-         actor_id,
+         %actor_struct{id: actor_id},
          changes,
          metadata,
          update_fun
@@ -278,10 +294,15 @@ defmodule Lightning.Auditing.Audit do
         event: event,
         item_id: item_id,
         actor_id: actor_id,
+        actor_type: extract_actor_type(actor_struct),
         changes: changes,
         metadata: metadata
       },
       update_fun
     )
+  end
+
+  defp extract_actor_type(struct_name) do
+    struct_name |> Module.split() |> List.last() |> Macro.underscore()
   end
 end